--- src/get_version.rs
@@ -1,37 +1,0 @@
-use ckb_build_info::Version;
-
-// Copy from https://github.com/nervosnetwork/ckb/blob/develop/src/main.rs
-pub(crate) fn get_version() -> Version {
-    let major = env!("CARGO_PKG_VERSION_MAJOR")
-        .parse::<u8>()
-        .expect("CARGO_PKG_VERSION_MAJOR parse success");
-    let minor = env!("CARGO_PKG_VERSION_MINOR")
-        .parse::<u8>()
-        .expect("CARGO_PKG_VERSION_MINOR parse success");
-    let patch = env!("CARGO_PKG_VERSION_PATCH")
-        .parse::<u16>()
-        .expect("CARGO_PKG_VERSION_PATCH parse success");
-    let dash_pre = {
-        let pre = env!("CARGO_PKG_VERSION_PRE");
-        if pre == "" {
-            pre.to_string()
-        } else {
-            "-".to_string() + pre
-        }
-    };
-
-    let commit_describe = option_env!("COMMIT_DESCRIBE").map(ToString::to_string);
-    #[cfg(docker)]
-    let commit_describe = commit_describe.map(|s| s.replace("-dirty", ""));
-    let commit_date = option_env!("COMMIT_DATE").map(ToString::to_string);
-    let code_name = None;
-    Version {
-        major,
-        minor,
-        patch,
-        dash_pre,
-        code_name,
-        commit_describe,
-        commit_date,
-    }
-}
